test(app-hooks): 拆分 useAppServerEvents 路由测试
